+++ bootique-jersey-client-instrumented/src/main/java/io/bootique/jersey/client/instrumented/ClientTimingFilter.java
@@ -2,18 +2,17 @@ package io.bootique.jersey.client.instrumented;
 
 import com.codahale.metrics.MetricRegistry;
 import com.codahale.metrics.Timer;
+import io.bootique.jersey.client.log.RequestLoggingFilter;
 import io.bootique.metrics.MetricNaming;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.ws.rs.client.ClientRequestContext;
-import javax.ws.rs.client.ClientRequestFilter;
 import javax.ws.rs.client.ClientResponseContext;
-import javax.ws.rs.client.ClientResponseFilter;
 import javax.ws.rs.ext.Provider;
 
 @Provider
-public class ClientTimingFilter implements ClientRequestFilter, ClientResponseFilter {
+public class ClientTimingFilter extends RequestLoggingFilter {
 
 	private static final Logger LOGGER = LoggerFactory.getLogger(ClientTimingFilter.class);
 	private static final String TIMER_PROPERTY = ClientTimingFilter.class.getName() + ".timer";
@@ -40,13 +39,14 @@ public class ClientTimingFilter implements ClientRequestFilter, ClientResponseFi
 
 	@Override
 	public void filter(ClientRequestContext requestContext, ClientResponseContext responseContext) {
+		StringBuilder logMessage = getResponseMessage(requestContext);
 		Timer.Context requestTimerContext = (Timer.Context) requestContext.getProperty(TIMER_PROPERTY);
 
 		// TODO: this timing does not take into account reading response
 		// content... May need to add additional interceptor for that.
 		long timeNanos = requestTimerContext.stop();
 
-		LOGGER.info("Client request finished. Status: {}, time: {} ms.", responseContext.getStatus(),
-				timeNanos / 1000000);
+		logMessage.append("time: ").append(timeNanos / 1000000).append(" ms.").append(" Status: ").append(responseContext.getStatus());
+		LOGGER.info(logMessage.toString());
 	}
 }