+++ RELEASE-NOTES.md
@@ -1,5 +1,5 @@
 ## 0.26
-
+* #25 Logging client requests
 * #31 Cleaning up APIs deprecated since <= 0.25
 * #32 Update Jackson to 2.9.5
 * #33 Metrics renaming to follow naming convention